+++ src/main/java/br/com/solutis/locadora/service/AbstractService.java
@@ -1,8 +1,9 @@
 package br.com.solutis.locadora.service;
 
+import br.com.solutis.locadora.model.dto.AbstractDto;
+
 import java.util.List;
 
-import br.com.solutis.locadora.model.entity.dto.AbstractDto;
 
 public abstract class AbstractService<T extends AbstractDto> {
     public abstract T findById(Long id);
